fix: :construction: quit stepper bottom buttons and act by clicks in steps
