+++ src/components/request-step/StepMid.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import { ContingencyForm } from '../form/ContingencyForm';
+
+interface Props {
+  next: () => void;
+}
+export const StepMid = ({ next }: Props) => {
+  return <ContingencyForm onSuccess={[next]} />;
+};
