+++ src/components/request-step/StepOne.tsx
@@ -2,20 +2,25 @@ import React from 'react';
 import { CardActionContainer } from '../card-action-container/CardActionContainer';
 import { CardAction } from '../card-action.tsx/CardAction';
 
-export const StepOne = () => {
+interface Props {
+  next: () => void;
+}
+export const StepOne = ({ next }: Props) => {
   return (
     <CardActionContainer>
       <CardAction
         urlAvatar="https://travellersworldwide.com/wp-content/uploads/2022/05/shutterstock_1938868960.png.webp"
         title="Vacations"
         text="7 days"
-        onClick={() => console.log('hola')}
+        onClick={next}
+        // onClick={() => action('vacation')}
       />
       <CardAction
         urlAvatar="https://travellersworldwide.com/wp-content/uploads/2022/05/shutterstock_1938868960.png.webp"
         title="Contingency"
         text="7 days"
-        onClick={() => console.log('hola')}
+        onClick={next}
+        // onClick={() => action('contingency')}
       />
       {/* <CardAction
         urlAvatar="https://travellersworldwide.com/wp-content/uploads/2022/05/shutterstock_1938868960.png.webp"
