+++ src/hooks/index.ts
@@ -1,1 +1,2 @@
+export { useStepper } from './useStepper';
 export { useModal } from './useModal';
